refactor: moves the small-variant tier into a `small` module
Moves omics-variation/src/variant.rs to small.rs and the variant/ typed
modules under small/, repointing intra-crate paths from crate::variant to
crate::small. The tier stays reachable at its historical crate-root paths
and the variant alias, and gains the new small paths. This is a pure move
with no behavior change.
