+++ omics-variation/src/small/mnv.rs
@@ -9,10 +9,10 @@ use omics_molecule::compound::Nucleotide;
 use omics_molecule::sequence;
 use omics_molecule::sequence::Sequence;
 
-use crate::variant::Alteration;
-use crate::variant::Kind;
-use crate::variant::KindError;
-use crate::variant::base_interval;
+use crate::small::Alteration;
+use crate::small::Kind;
+use crate::small::KindError;
+use crate::small::base_interval;
 
 /// A multi-nucleotide variant, an equal-length substitution of two or more
 /// bases.
